do hard deletes in room DO
