+++ apps/dotcom/sync-worker/src/TLUserDurableObject.ts
@@ -2,7 +2,6 @@ import {
 	DB,
 	isColumnMutable,
 	MAX_NUMBER_OF_FILES,
-	ROOM_PREFIX,
 	TlaFile,
 	TlaFilePartial,
 	TlaFileState,
@@ -26,7 +25,7 @@ import { getR2KeyForRoom } from './r2'
 import { Analytics, Environment, TLUserDurableObjectEvent } from './types'
 import { UserDataSyncer, ZReplicationEvent } from './UserDataSyncer'
 import { EventData, writeDataPoint } from './utils/analytics'
-import { getReplicator, getRoomDurableObject } from './utils/durableObjects'
+import { getRoomDurableObject } from './utils/durableObjects'
 import { isRateLimited } from './utils/rateLimit'
 import { retryOnConnectionFailure } from './utils/retryOnConnectionFailure'
 
@@ -295,7 +294,11 @@ export class TLUserDurableObject extends DurableObject<Environment> {
 				if (!file) {
 					// The user might not have access to this file yet, because they just followed a link
 					// let's allow them to create a file state for it if it exists and is shared.
-					file = (await getReplicator(this.env).getFileRecord(nextFileState.fileId)) ?? undefined
+					file = await this.db
+						.selectFrom('file')
+						.selectAll()
+						.where('id', '=', nextFileState.fileId)
+						.executeTakeFirst()
 				}
 				if (!file) {
 					throw new ZMutationError(ZErrorCode.bad_request, `File not found ${nextFileState.fileId}`)
@@ -428,10 +431,6 @@ export class TLUserDurableObject extends DurableObject<Environment> {
 							const { id } = update.row as any
 							await tx.deleteFrom(update.table).where('id', '=', id).execute()
 						}
-						if (update.table === 'file') {
-							const { id } = update.row as TlaFile
-							await this.deleteFileStuff(id)
-						}
 						break
 				}
 				this.cache.store.updateOptimisticData([update], msg.mutationId)
@@ -532,37 +531,6 @@ export class TLUserDurableObject extends DurableObject<Environment> {
 	}
 
 	/* --------------  */
-
-	private async deleteFileStuff(id: string) {
-		const fileRecord = await getReplicator(this.env).getFileRecord(id)
-		const room = this.env.TLDR_DOC.get(this.env.TLDR_DOC.idFromName(`/${ROOM_PREFIX}/${id}`))
-		await room.appFileRecordDidDelete()
-		if (!fileRecord) {
-			throw new Error('file record not found')
-		}
-		const publishedSlug = fileRecord.publishedSlug
-
-		// Create a new slug for the published room
-		await this.env.SNAPSHOT_SLUG_TO_PARENT_SLUG.delete(publishedSlug)
-
-		// remove published files
-		const publishedPrefixKey = getR2KeyForRoom({
-			slug: `${id}/${publishedSlug}`,
-			isApp: true,
-		})
-		const publishedHistory = await listAllObjectKeys(this.env.ROOM_SNAPSHOTS, publishedPrefixKey)
-		if (publishedHistory.length > 0) {
-			await this.env.ROOM_SNAPSHOTS.delete(publishedHistory)
-		}
-		// remove edit history
-		const r2Key = getR2KeyForRoom({ slug: id, isApp: true })
-		const editHistory = await listAllObjectKeys(this.env.ROOMS_HISTORY_EPHEMERAL, r2Key)
-		if (editHistory.length > 0) {
-			await this.env.ROOMS_HISTORY_EPHEMERAL.delete(editHistory)
-		}
-		// remove main file
-		await this.env.ROOMS.delete(r2Key)
-	}
 
 	private async publishSnapshot(file: TlaFile) {
 		if (file.ownerId !== this.userId) {
@@ -675,19 +643,6 @@ export class TLUserDurableObject extends DurableObject<Environment> {
 		}
 		return cache.store.getCommittedData()
 	}
-}
-
-async function listAllObjectKeys(bucket: R2Bucket, prefix: string): Promise<string[]> {
-	const keys: string[] = []
-	let cursor: string | undefined
-
-	do {
-		const result = await bucket.list({ prefix, cursor })
-		keys.push(...result.objects.map((o) => o.key))
-		cursor = result.truncated ? result.cursor : undefined
-	} while (cursor)
-
-	return keys
 }
 
 class ZMutationError extends Error {
